# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PhishGuard API",
        "description": "API untuk deteksi dan pelaporan phishing",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://be-phisguard-production.up.railway.app/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/report": {
            "post": {
                "tags": [
                    "Report"
                ],
                "summary": "Submit phishing report",
                "operationId": "4b64511eec46a94d687d3f3f174bc9c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "channel_chat",
                                    "sender_account",
                                    "chat_text"
                                ],
                                "properties": {
                                    "channel_chat": {
                                        "type": "string",
                                        "example": "WhatsApp"
                                    },
                                    "sender_account": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "chat_text": {
                                        "type": "string",
                                        "example": "Klik link hadiah ini"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "http://phishing.com"
                                    },
                                    "reporter_name": {
                                        "type": "string",
                                        "example": "Pijen"
                                    },
                                    "region": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    },
                                    "interaksi": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "incident_summary": {
                                        "type": "string",
                                        "example": "Pengguna menerima pesan phishing dan mengklik link"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Report submitted successfully"
                    }
                }
            }
        },
...
        "/report/{id}": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Get report detail",
                "operationId": "d6fa9f7e95050abe887fd35e855a9dcf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report detail"
                    },
                    "404": {
                        "description": "Report not found"
                    }
                }
            }
        },
...
    },
...
}